requirements.txt
Appium-Python-Client==0.41
pytest==4.4.2
allure-pytest==2.6.2